Redirect to sign up page if unauthenticated user tries to vote a post

+++ frontend/components/posts/PostVotes.tsx
@@ -6,13 +6,15 @@ import { Post } from '../../src/types/Post'
 import IconButton from '@material-ui/core/IconButton'
 import gql from 'graphql-tag'
 import { useMutation } from '@apollo/react-hooks'
+import Router from 'next/router'
 
 interface Props {
   post: Post
+  userId: string | undefined
 }
 
 export const PostVotes = (props: Props) => {
-  const { post } = props
+  const { post, userId } = props
   const [createPostVote] = useMutation(CREATE_POST_VOTE_MUTATION)
   const [deletePostVote] = useMutation(DELETE_POST_VOTE_MUTATION)
   const [viewerVote, setViewerVote] = useState(post.viewerVote ? (post.viewerVote.value === 'POSITIVE_1' ? 1 : -1) : 0)
@@ -38,6 +40,11 @@ export const PostVotes = (props: Props) => {
 
   const handleVote = async (e: React.MouseEvent<HTMLButtonElement>, value: 1 | -1) => {
     e.preventDefault()
+
+    if (!userId) {
+      Router.push('/register')
+      return
+    }
 
     if (value === viewerVote) {
       await handleDeleteVote(true)
